Remove test for Connection #finalize
This hurts bad to do, and I'm open to any suggestion for how to
do this test in a way that is deterministic.

Whenever #finalize is called on a user-level Session object,
we check that the session was properly closed. If it was not
closed, we generate an error and close the session.

The test for this would create a Session object with no references,
and then run tens of thousands of iterations of System.gc(),
testing for #close() being invoked. This seems to have worked
deterministically on Java 8, but on Java 7 it is flaky.

I don't think testing for Java 7 in a JUnit assumption is
a good solution, and instead I'm opting for removing the test,
with the argument that this has been manually verified to work,
it is a best-effort piece of code that in production usage is
allowed to fail (since we can never guarante the finalizer will
invoke us).

If we drop java 7 support, we could look at reverting this commit.