Change ResultCursor to Iterator<Result>
When we first designed the result cursor, the intention was to build a
low-overhead API that didn't require allocating lots of objects to move
through the result.

In testing this with developers, there's been a fair bit of negative
feedback and, more concerning - misunderstanding the model. Many
developers are not used to the Cursor pattern, and don't untuitively
imagine it as "moving a view" over the result. This leads to dissonance,
which leads to frustration, which leads to suffering, which leads to the
dark side.

Tipping the scale is a series of experiments with escape analysis we
did. It turns out that escape analysis is powerful enough to avoid heap
allocation even for nested calls like:

    result.next().get("blah").get(0).asInt();

The construct above can be allocated entirely on the stack, even if each
method but the last returns a newly allocated object. This has been
verified in experiments on HotSpot 25.74-b02, Java 8.

Notably, the current implementation does not allow this, since Record
is not allocated directly in the `next` method - but this is of no
significance, because this was the case for the Cursor API as well. The
point is that the API *allows* us to write heap-allocation-free impl in
the future.

This means that we can use an Iterator-based approach, which is MUCH
more intuitive to end users. It also means we can do away with a
significant amount of edge cases ("this method only works if the cursor
is in so-and-so position") and a long slew of methods put in place to
avoid allocation (eg. `single(<fieldName>)` and friends).

Finally, it means there is no longer a distinction between 'retaining'
an individual record and 'viewing' a record. A record is always an
immutable copy, if you store it somewhere it works like any other
object. If you don't, escape analysis will allow allocating it on the
stack.