Final restructuring to help discoverability of API.
Move interfaces that are not needed to get going into dedicated
namespaces - summary, util and value - collapse unused interfaces,
and finally remove Identity type.

This creates a root API namespace which should be significantly easier
to get started with, and through which users will discover the value
types and, if they need them, the summary types.