Refactor routing logic to better handle failures
Separate out different parts of the logic into different classes:
 * When to update the view of the cluster, and the maintenance of the
   set of servers in the cluster:
   - LoadBalancer
 * The ability to retrieve a list of servers with roles
   (ClusterComposition) from a server:
   - ClusterComposition.Provider
 * Round Robin cycling through a set of addresses (maintaining the
   order/position on addition/removal of members in the set):
  - RoundRobinAddressSet

The new logic distinguishes being unable to connect to the routing
servers and maintains the server in the list in this case, under the
assumption that the server is only temporarily unavailable. The
attempt to contact routing servers only fail after a number of failed
attempts with an exponential back off in between.

These changes introduce event based testing for some of the components,
this makes it easier to test for the behaviour of units, rather than
asserting on their state, while the explicitness of it makes it easier
to follow than using mocks.