Remove ping on Session#close()
Previously ping query (`RETURN 1`) was executed on each `Session#close()`.
This was intended to verify the underlying connection and decide if it should
be returned to the pool or not.

Running such ping is not quite necessary because we already execute `RESET` and
`SYNC` before that. Combination of `RESET` and `SYNC` guarantees to send at
least one message and thus validates the connection. Also query `RETURN 1`
made database start transaction and possibly do planning, this work seem
unnecessary.

This commit removes ping and adds tests to verify that we actually send
messages to socket during pooled connection validation.