Fail-fast interactions with closed driver
It was previously possible to use closed driver instance, obtain sessions from
it and run queries. This was seen to cause problems in multi-threaded scenarios
when some worker threads try to use driver, which was concurrently closed by
other worker.

This commit makes driver track if it is closed or not and throw exceptions if
on every interaction when closed. Thrown exception is `IllegalStateException`.
Actuall `#close()` method can be called multiple times and is a no-op for
already closed driver.