Better internal error handling
Introduced a few internal errors for tracking errors internally as checked errors.
Before the internal errors go outside (Driver, Session, Transaction, Result), the errors will be casted to a public error.

Each method is marked with the specific errors that it is going to throw for now. This could be put into one general error once we agree on the internal/external error mappings.