Better handle async close in SocketConnectionPool
This pool can be used from multiple threads. It is possible for `#close()` to
be called concurrently with `#acquire()`. Previously we did not guard against
this situation and it was possible to acquire connections from the closed pool.
It could also happen that `#close()` does not actually close all connection
queues because iteration over them is not atomic.

This commit makes `SocketConnectionPool` track if it is closed or not. It also
introduces `Connector` abstraction (which is merely a connection factory) so
that `SocketConnectionPool` can be easily tested.