Close connection if init failed
Init message is send every time we establish a new socket connection. This is
the very first thing connection does. It is possible for the init to fail
(for example when credentials are wrong). Previously socket connection was not
closed after such init failure.

This commit changes `SocketConnector` to close connection on any init failure.