Close connection pool if driver creation fails
Connection pool (`SocketConnectionPool`) is created before drivers and was
never closed if driver creation failed. This was especially possible with
`RoutingDriver` which tries to build routing table in constructor.

This commit adds closing of the connection pool when creation of driver fails.
It also extracts driver creation logic into an internal class `DriverFactory`
to make it testable.