Remove terminated connection queues from the pool
It was previously possible for empty/terminated connection queues to stay in
the `SocketConnectionPool` when is was closed concurrently with sessions being
acquired. This was not critical because `SocketConnectionPool` is only closed
when the whole driver is closed but resulted in a flaky test.

This commit adds removal of terminated connection queues.