Configurable test-on-borrow for pooled connections
Connections are pooled for both direct and routing driver. Each acquired
session contains a socket connection which was acquired from the pool.
It is possible for an idle pooled connection to become invalid while just
resting in the pool. This could happen when connection is terminated after
some timeout by a load balancer or some other network facility.

This commit introduces a configurable connection liveness check timeout.
So freshly acquired connection, that has been idle in the pool for more
than configurable timeout, will be validated by sending a RESET message.
If it appears to be broken acquisition will retry until a valid
connection is found.