First step to improve internal error handling
In this PR, we use PooledConnection as a cut of all internal checked errors and external runtime exceptions. All internal errors will be gathered in PooledConnection and they have to be cast to external Neo4j errors before leaving this interface. This PR enables us to clean all runtime errors bellow PooledConnection, while keep the serface the same as before.