Wait for bookmark when starting a transaction
It is possible to specify bookmark when starting a transaction. First `BEGIN`
statement would then contain the given bookmark in params. Database will wait
when processing such `BEGIN` for the state described by the bookmark to appear.

Sending `BEGIN` and bookmark was previously lazy. So it was possible to start
a transaction with an invalid/unreachable bookmark. Error would in such case
pop up only after the next sync (commit or consume) which is too late and can
be quite confusing.

This commit makes transaction eagerly send `BEGIN` and `PULL_ALL` messages
when non-null bookmark is given. It makes `session#beginTransaction("bookmark")`
call blocking but the API less surprising.