Decouple session from connection
Previously session was bound to a connection and used this same connection
throughout its lifetime. This means that there was no load-balancing within
single session and that connection could have been claimed by the session
for a long time without actually being used.

This commit makes session obtain new connections when needed and return them
to the pool as soon as possible. New connection is obtained for:
 * each `Session#run()`
 * each `Session#beginTransaction()` and created transaction holds same
 connection throughout its lifetime
Connection is returned back to the pool when either:
 * result is fully read into memory
 * transaction is closed
 * subsequent `Session#run()` is invoked

Introduced new interface `ConnectionHandler` to manage connections around
session, result and transaction. `NetworkSession` is the only implementation
of this interface.
Also added new `DriverClosedException` thrown on every interaction with a
closed driver.