Close connection after result reading failure
When result fetching fails (cypher runtime error or network error) exception
is thrown but it is still possible to maintain a reference to the result
object. It'll just be empty after failure. It is not possible to fetch any
other records after failure and the underlying pooled connection can be
closed and returned to the pool.

This commit makes result return its connection to the pool on failure. This
connection (if healthy) can be used by subsequent sessions. Also added tests
for connection handling.