Allow bookmark in `Driver#session()`
This commit adds new API methods to create session with the specified bookmark.
This means first transaction will ensure that the server, it is connected to,
is at least up to the database version denoted by the bookmark.

Following new methods are introduced in the `Driver` interface:
 * `#session(String bookmark)`
 * `#session(AccessMode mode, String bookmark)`

Bookmark is a session related concept and this commit adds symmetry with
regards to `Session#lastBookmark()` method.