Simplify retry logic
It was previously represented with two interfaces - `RetryLogic` and
`RetryDecision`. First containing the actual code to decide if retry
should happen and second holding the state between retries. Approach
works but it is unnecessarily complex.

This commit simplifies rety logic to a single interface `RetryLogic`
with a single method that takes a function to execute and performs
all needed retries.