Multiple bookmarks
Previously it was possible to only supply a single bookmark when creating
a new session. However there is a use-case to be able to supply multiple
bookmarks when multiple worker threads execute write queries and then
reader should be able to observe all those writes. To achieve this driver
now exposed session creation methods that take an iterable of bookmarks.

It was chosen to add additional `Driver#session()` overloads with
`Iterable<String>` instead of extending single-bookmark methods with
varargs because:
 * this maintains backwards compatibility for pre-existing methods
 * addition of varargs would make API look strange because call or
   `Driver#session()` and call of `Driver#session(String... bookmarks)`
   without params is essentially the same thing
 * invocation of vararg methods with single `null` argument can be confusing
 * varargs are more suitable for manual params listing

Driver will now send:
```
{
    bookmark: "max",
    bookmarks: ["one", "two", "max"]
}
```
instead of simple:
```
{
    bookmark: "max"
}
```
this is done to maintain backwards compatibility with databases that
only support single bookmark. It forces driver to parse and compare
bookmarks which violates the fact that bookmarks are opaque. This is
done only to maintain backwards compatibility and should not be copied.
Code doing this will eventually be removed.