Guard IPv6 test with an assumption
Commit adds a database version assertion to IPv6 direct driver test.
It is needed because only 3.1+ neo4j versions support
"dbms.connectors.default_listen_address" configuration option which
we need to set to "::" for this test. Test got broken when direct
driver started to verify connectivity on creation. Previously it
simply accepted IPv6 address but did not try to connect.