Added ability to enforce max connection lifetime
Driver keeps idle socket connections in a connection pool. These
connections can get invalidated while resting idle in the pool.
They can be killed by network equipment like load balancer, proxy or
firewall. It is also safer to refresh connections once in a while so
that database can renew corresponding resources.

This commit adds `maxConnectionLifetime` setting and makes driver
close too old connections. Checking and closing happens in an
application thread that tries to acquire a connection. Feature can
be used on it's own or in combination with
`connectionLivenessCheckTimeout` to guarantee validity of acquired
connections.