Simplify FixedRetryLogic
This is a test-only implementation of `RetryLogic` interface that
retries up to a specified amount of times. It previously had a
separate implementation which was basically a simplified version
of `ExponentialBackoffRetryLogic`. It is problematic for following
`retryAsync` functionality because implementation will not be that
simple. This commit makes `FixedRetryLogic` extend
`ExponentialBackoffRetryLogic` and override just a single method.
So that they can share complicated retrying code.