Simplify async result cursor record fetching
Commit replaces `Response<Boolean> #fetchAsync()` + `Record #current()`
in `StatementResultCursor` with single `Response<Record> #nextAsync()`.
This simplifies API and makes it more convenient to consume records via
future chaining and callbacks. `Response` is already a container for
record, it can't contain anything else. So it makes sense to return
`null` as an end of records stream marker and not use separate method
for this.

Also added `#peekAsync()` method.