Session#closeAsync() waits result to be fetched
This commit aligns behaviour of `Session#close()` and `Session#closeAsync()`
so that later one also waits for the latest result to be fully fetched.
It makes possible to consume result after session is closed and allows
propagation of query errors in `#closeAsync()`. Errors are propagated
only if they were not consumed by reading the result cursor.