Session#reset() releases connection last
Previously it tried to first release connection and them mark existing
transaction as terminated. This could've been problematic because it
allowed transaction to be used on the client after RESET, which clears
connection state on the server.

This commit makes `Session#reset()` first mark existing transaction as
terminated and only then release the connection. Releasing connection
implies RESET.