Fixed inheriting of output from stub server
It got broken because of maven-surefire-plugin update, which now prohibits
`ProcessBuilder#inheritIO()` and only allows writing to stdout. Output
with `#inheritIO()` did not go to the command line but instead got written
to a separate temporary file. Warning "Corrupted stdin stream in forked JVM 1"
has also been generated.

Seems like maven-surefire-plugin prohibits interactions with the output
stream because that's how it talks to the forked JVM. By default
"forkCount" is 1 so such JVM exists in our case.

See:
    http://maven.apache.org/surefire/maven-surefire-plugin/faq.html#corruptedstream
    https://issues.apache.org/jira/browse/SUREFIRE-1359