Fixed auto-read problem with result summary
All result stream should be buffered/consumed when result summary is
requested or when session is closed. This means underlying channel
should read everything that's available. It did not do so when summary
was requested after number of buffered records exceeded high watermark.

This commit makes fetching of summary or failure enable auto-read on
the underlying channel. So SUCCESS and FAILURE messages will be read
without waiting for all records to be consumed.