Cleanup of CompletionException handling
Exceptions can get wrapped in `CompletionException` when propagated
through `CompletionStage#whenComplete()` and `CompletionStage#handle()`.
It is also needed to wrap throwable in a `CompletionException` before
throwing it from a `CompletionStage` chain because checked exceptions
are not allowed there.

This commit cleans up helpers to unwrap `CompletionException` and
convert throwable to `CompletionException`.