Optimized `#listAsync()`
Method `InternalStatementResultCursor#listAsync()` returns a future
containing list of all records. Records can potentially be transformed
using the `#listAsync(Function<Record, T>)` overload. Implementation
used to be built on top of `#nextAsync()` and invoked it recursively
in `ForkJoinPool.commonPool` workers. This resulted in quite a lot of
garbage generation because every record needed to be wrapped in a
`CompletableFuture`. Also execution in the separate threads resulted
in increased CPU load.

`PullAllResponseHandler` is responsible for buffering of incoming
records in a queue. This commit makes `#listAsync()` implementation
wait for final PULL_ALL message and then transform the queue into list.
When FAILURE is received for PULL_ALL resulting future is failed. When
SUCCESS is received for PULL_ALL given transformation function is
applied to the queue of records and resulting list is returned. This
results in less garbage being generated because no intermediate
`CompletableFuture`s are created.