Optimize `StatementResultCursor#consumeAsync()`
Previously it was implemented using `#forEachAsync()` and a no-op
consumer. This resulted in every record being added to the queue of
incoming records, then wrapped in a `CompletableFuture` and only then
thrown away. So it did a lot more work than needed.

This commit optimizes `#consumeAsync()` and `StatementResult#consume()`
that builds on top of it. Now all incoming values will not be
transformed into records and will be dropped right away. So no queueing
and `CompletableFuture`s will be involved.