Better sanitize auth token parameters
Previously it was possible to send `null` as username, password, etc.
which resulted in a database error saying that parameter value is
missing. It was also possible to send `null` realm, which is
essentially no realm.

This commit adds `null` validation in the driver code so that it fails
as early as possible. It also makes driver not send the realm
when it's `null`.