Fixed error when connection killed during tx
Connection error can happen while transaction is executing. Closing
transaction after such error should not perform rollback because
connection is dead and database has cleaned up all resources.

Previously close operation tried to perform rollback and failed with
same `ServiceUnavailable` exception as the unsuccessful `#run()`. Error
thrown from `#close()` has to be added as a suppressed error to the one
thrown from `#run()`, when used in try-with-resources block. So code
attempted to add error to itself as suppressed. This resulted in an
`IllegalArgumentException`.

This commit fixes the problem by making `#close()` not perform a
rollback after a fatal connection error.