Correctly prohibit multiple transactions in session
Session can only have one active (not closed) transaction at a time.
Exception is thrown when starting a transaction on a session that
already has an open one.

Situation when transaction is started multiple times on a session with
an open transaction was not handled correctly. First failure overrode
the reference to the last transaction kept in the session. Later
attempts resulted in a different error saying that there already
exists an open connection.

This commit fixes the problem by making session keep current transaction
reference even if new transaction failed to begin. So callers will
receive consistent errors.