Make `Session#reset()` only send RESET message
Instead of sending RESET message and returning connection back to the
pool. Queries & transactions should still be in control of the
connection lifecycle even after reset. Previous behaviour was
problematic because reset might race with query execution and make
it try to use connection that has been returned to the pool.

Also moved all RESET tests to a dedicated `SessionResetIT`. Made
explicit transaction remain open after termination so that user code
still needs to close it to return connection to the pool.