Close open transaction when session is closed
Instead of rolling it back. This was the previous behaviour that got
unintentionally changed in 1.5.0 version of the driver. It makes more
sense to close the transaction (which means commit or rollback based
on it's state) than just silently rollback. So driver will adhere to
`Transaction#success()` or `Transaction#failure()` markers when
closing session with an open transaction.