Reduce code duplication in statement runners
Interface `StatementRunner` defines methods for running cypher queries
with different forms of parameters. It is implemented by both session
and transaction. Both used to implement overloads of `#run()` and
`#runAsync()` in a similar way.

This commit adds an abstract class for session and transaction so that
they only need to implement a single `#run(Statement)` and
`#runAsync(Statement)`.