Handle nanosecond normalization in IsoDuration
Database normalizes nanoseconds to be in range from 0 to 999_999_999.
`IsoDuration` previously did no calculations in it's `#toString()`
and returned incorrect result.

This commit makes it return correct string representation despite
normalized second and nanosecond components. It will also do
normalization in the constructor using
`java.time.Duration#ofSeconds(long, long)` in order to return correct
strings for durations created on the client side and not returned
from the database.