Support hostname verification for secure connections
Hostname verification makes client verify that server it is trying to
connect to matches server specified in the returned certificate. This
is a way to protect from man in the middle attack and make sure client
is talking to the correct server. Hostname will be compared to all
entries in `subjectAltName` field of the certificate.

This commit makes driver support configurable hostname verification. It
can be configured together with the trust strategy:

```
Config config = Config.build()
    .withTrustStrategy(trustAllCertificates().withHostnameVerification())
    .toConfig();
```

Method `#withHostnameVerification()` is available for all trust
strategies. Verification is turned off by default.