Improve handling of RESET and ACK_FAILURE errors
This commit makes driver close the network channel when either `RESET`
or `ACK_FAILURE` receive an error response. Message `RESET` is sent
before release of the channel back to the channel pool. It might fail
if database experiences a fatal error (like `OutOfMemoryError`) or a
network failure occurs. In case of such errors channel will be closed
and returned to the pool, which will drop it later. Message
`ACK_FAILURE` is sent to acknowledge a `FAILURE` response. Failure
to do so is unrecoverable and should result in a connection
being dropped.