Fix handling of codec exceptions without cause
Driver tries to extract cause of every `CodecException` received in
`#exceptionCaught(Throwable)` netty callbacks. It did not handle
cases when the cause is null, which resulted in an attempt to
exceptionally complete a `CompletableFuture` with null.

This commit fixes the problem by making handlers extract the exception
cause only if it is defined and propagate the original error otherwise.