Configurable server address resolver
This commit makes it possible to configure a resolver function used by
the routing driver. Such function is used during the initial discovery
and when all known routers have failed. Driver already had an internal
facility like this which performed a DNS lookup for the hostname of
the initial address. This remains the default. Users are now able to
provide a custom resolver in the config.