Improve neo4j feature detection in tests
Some tests are only relevant for all neo4j versions which have a
particular feature. This was previously checked using assumptions
and neo4j version comparisons. This commit adds a JUnit 5 test
execution condition which is based on neo4j feature. Every test
can be enabled/disabled based on presence/absence of a feature
specified in an annotation.