Wait for all errors to arrive before closing tx
Results of all queries need to be consumed before transaction is
committed or rolled back. This is done by `ResultCursorsHolder`
which keeps all cursors and queries them for errors. Previously,
it only waited for the first failure to arrive and there might've
been some more queries receiving results.

This commit fixes the problem by making the `ResultCursorsHolder`
properly await for all result cursors to be consumed. It does not do
blocking await but structures the result future to only be completed
when all results have arrived.