Allow conversion of statement result to stream
Queries executed via blocking API return `StatementResult` which is an
iterable of records. This commit adds a convenience method
`StatementResult#stream()` to convert result to a stream. It should
be convenient to use this method when processing of returned records
is required.

Example:

```
List<Integer> list = session.run("UNWIND range(1, 100) AS x RETURN x")
                            .stream()
                            .map(record -> record.get("x"))
                            .map(Value::asInt)
                            .collect(toList());
```