Ensure a single handler manages channel's auto-read
`PullAllResponseHandler` is responsible for receiving a stream of
records. Stream might be very large and handler tries to not fetch
if fully into memory. It applies a network-level backpressure by
disabling/enabling auto-read property of the underlying Netty channel.
This property tells channel to not read from the network. Auto-read is
disabled when number of queued records goes beyond the threshold.

Management of auto-read turned out to be problematic for nested
queries within a single transaction. Nested queries result in
multiple `PullAllResponseHandler`s being added to the queue of
handlers. They will try to manage auto-read concurrently and can
sometimes disable it completely. Callers would then be blocked
and unable to proceed. This is not a problem for `Session#run()`
because every such call ensures there was a logical SYNC and the
previous query has completed and its result is buffered.

This commit fixes a problem by making only a single installed handler
manage the channel's auto-read property. Making sure there only exists
a single such handler is the responsibility of
`InboundMessageDispatcher`. Auto-read is enabled when new auto-read
managing handler is installed. Previous such handler is disabled.