Disable the removed auto-read managing handler
To make sure it does not influence the auto-read status of the channel.
Also restore default value of auto-read for channel when such handler
is removed.

This change is needed to avoid removing a handler that has just
disabled auto-read. This is possible when handler receives a RECORD
that causes it to disable auto-read and then receives a SUCCESS
message, which causes it to be dequeued with auto-read being disabled.