Change how BoltServerAddress equality works
BoltServerAddress was checking equality on  `Hostname`, `Resolved IP` and `Port` triplet and DnsResolver was causing loss of original `Hostname`.

This commit removed public access to resolved IP Address component of BoltServerAddress and removed it from the equality check, thus resulting in same `hostname:port` pairs equal whatever IP address it is resolved to.