Unified the input and output value of database name.
The database name default to `null` if the server does not support multi-databases when the database name is returned in result summary.
However `null` is a forbidden value for database in session parameters. Then for user who want to connect to default db but do not want ot omit invoke `withDatabase` method on session parameter, they have to use empty string to work aroud our API design.

This PR allows `null` as a valid database name for session parameters, while disallows empty string as it is an invalid database name to users.